Auto-throw exceptions when resuming tasks
Before, whenever a task waited for a blocker it had to call
tasks.check(blocker) afterwards to check whether the blocker had an exception.
Now, taking advantage of generator.throw in Python 2.6, the exception is raised
directly from the yield statement.

When upgrading existing code, it is always safe to transform:

	yield foo
	...

into

	try:
		yield foo
	except:
		pass
	...

Normally, however, you can simply eliminate the check, so:

	yield foo
	tasks.check(foo)

becomes just:

	yield foo